backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
google-genai
python-dotenv
python-multipart
numpy
